Refactor and fix test
